requirements.txt
minio==7.2.12
python-dotenv==1.0.0
pyspark==3.5.3
numpy
kafka-python
Faker